+++ libs/backend/agents/personal/personas/main/src/index.ts
@@ -3,7 +3,7 @@ export { __CreatePersonaOnboardingRouter } from "./http/persona-onboarding.route
 export { _CreatePersonaOnboardingRouter } from "./http/prisma-persona-onboarding.router";
 export { _PersonaOnboardingOpenapiPaths } from "./http/openapi";
 export type { PersonaOnboardingCaller, PersonaOnboardingClock, PersonaOnboardingRouterDependencies, PersonaOnboardingWorkflowPort } from "./http/persona-onboarding.router.types";
-export type { PersonaAgentRevisionSelectionFactory } from "./http/prisma-persona-onboarding.router.types";
+export type { PersonaAgentRevisionSelectionFactory } from "./profile/prisma-persona-persistence-composition.types";
 export { PersonaAgentRevisionSelectionStatuses } from "./approval/persona-authority.types";
 export type { PersonaAgentRevisionSelectionPort, SelectApprovedPersonaForPersonalAgentCommand, SelectApprovedPersonaForPersonalAgentResult } from "./approval/persona-authority.types";
 export { _CreatePersonaWorkflowEvidenceRepository } from "./profile/prisma-persona-workflow-evidence";
