+++ libs/backend/agents/personal/personas/main/src/profile/prisma-persona-persistence-composition.types.ts
@@ -1,10 +1,8 @@
-import type { Prisma } from "@prisma/client";
-
 import type { PersonaAgentRevisionSelectionPort } from "../approval/persona-authority.types";
 
 /** App-owned factory that binds persona approval to agent-service revision selection. */
-export interface PersonaAgentRevisionSelectionFactory
+export interface PersonaAgentRevisionSelectionFactory<Transaction = unknown>
 {
 	/** Creates the narrow selection port over the persona unit of work's transaction. */
-	create(transaction: Prisma.TransactionClient): PersonaAgentRevisionSelectionPort;
+	create(transaction: Transaction): PersonaAgentRevisionSelectionPort;
 }
