+++ libs/backend/server/agents/agent-services/main/src/__tests__/prisma-personal-agent-bootstrap-repository.test.ts
@@ -70,55 +70,6 @@ describe("Prisma personal-agent bootstrap repository", function _Suite()
 		expect(transaction.personaRevision.findUnique).not.toHaveBeenCalled();
 	});
 
-	it("creates, publishes, activates, and audits one initial personal agent", async function _CreatesReadyAgent()
-	{
-		const transaction = _Transaction();
-		const result = await _Repository(transaction).ensureReady(_COMMAND);
-
-		expect(result).toEqual({ status: PersonalAgentBootstrapStatuses.Ready, agentServiceId: _COMMAND.onboardingId, agentRevisionId: "revision-a", created: true, revised: false });
-		expect(transaction.agentService.create).toHaveBeenCalledWith({
-			data: expect.objectContaining({
-				id: _COMMAND.onboardingId,
-				siloId: _COMMAND.siloId,
-				kind: "Personal",
-				state: "Draft",
-				name: "The Commander",
-				workloadProfile: "personal-default",
-			}),
-		});
-		expect(transaction.agentRevision.create).toHaveBeenCalledWith({
-			data: expect.objectContaining({
-				revision: 1,
-				promptPolicyVersion: INITIAL_PERSONAL_AGENT_POLICY.promptPolicyVersion,
-				personaRevisionId: _COMMAND.onboardingPersonaRevisionId,
-				budget: { maxTurns: 64, maxTokens: 256_000, maxDurationMs: 3_600_000 },
-				modelDefinition: { connect: { id: "tenant-default" } },
-			}),
-			include: expect.any(Object),
-		});
-		expect(transaction.agentRevision.update).toHaveBeenCalledWith({ where: { id: "revision-a" }, data: { state: "Published", publishedAt: _COMMAND.provisionedAt } });
-		expect(transaction.agentService.update).toHaveBeenCalledWith({ where: { id: _COMMAND.onboardingId }, data: { state: "Active", activeRevisionId: "revision-a", updatedAt: _COMMAND.provisionedAt } });
-		expect(transaction.auditDecision.create).toHaveBeenCalledOnce();
-	});
-
-	it("prefers one tenant default without consulting the global fallback", async function _TenantDefaultWins()
-	{
-		const transaction = _Transaction();
-		await _Repository(transaction).ensureReady(_COMMAND);
-
-		expect(transaction.modelDefinition.findMany).toHaveBeenCalledOnce();
-		expect(transaction.modelDefinition.findMany).toHaveBeenCalledWith(expect.objectContaining({ where: { scope: "ClusterTenant", clusterTenant: _COMMAND.siloId, isDefault: true } }));
-	});
-
-	it("uses one global default when the silo has no default", async function _GlobalFallback()
-	{
-		const transaction = _Transaction();
-		transaction.modelDefinition.findMany.mockReset().mockResolvedValueOnce([]).mockResolvedValueOnce([{ id: "global-default" }]);
-
-		await expect(_Repository(transaction).ensureReady(_COMMAND)).resolves.toMatchObject({ status: PersonalAgentBootstrapStatuses.Ready });
-		expect(transaction.agentRevision.create).toHaveBeenCalledWith(expect.objectContaining({ data: expect.objectContaining({ modelDefinition: { connect: { id: "global-default" } } }) }));
-	});
-
 	it("returns the deterministic ready winner without duplicating writes", async function _IdempotentWinner()
 	{
 		const transaction = _Transaction();
@@ -197,29 +148,4 @@ describe("Prisma personal-agent bootstrap repository", function _Suite()
 		expect(transaction.agentService.create).not.toHaveBeenCalled();
 	});
 
-	it("fails closed when a model scope has multiple defaults", async function _AmbiguousDefault()
-	{
-		const transaction = _Transaction();
-		transaction.modelDefinition.findMany.mockReset().mockResolvedValueOnce([{ id: "model-a" }, { id: "model-b" }]);
-
-		await expect(_Repository(transaction).ensureReady(_COMMAND)).resolves.toEqual({ status: PersonalAgentBootstrapStatuses.Denied, reason: PersonalAgentBootstrapDenialReasons.DefaultModelAmbiguous });
-		expect(transaction.agentService.create).not.toHaveBeenCalled();
-	});
-
-	it("fails closed when neither scope has a default model", async function _MissingDefault()
-	{
-		const transaction = _Transaction();
-		transaction.modelDefinition.findMany.mockReset().mockResolvedValueOnce([]).mockResolvedValueOnce([]);
-
-		await expect(_Repository(transaction).ensureReady(_COMMAND)).resolves.toEqual({ status: PersonalAgentBootstrapStatuses.Denied, reason: PersonalAgentBootstrapDenialReasons.DefaultModelUnavailable });
-		expect(transaction.agentService.create).not.toHaveBeenCalled();
-	});
-
-	it("propagates audit failure so the onboarding transaction can roll everything back", async function _AuditFailure()
-	{
-		const transaction = _Transaction();
-		transaction.auditDecision.create.mockRejectedValue(new Error("audit unavailable"));
-
-		await expect(_Repository(transaction).ensureReady(_COMMAND)).rejects.toThrow("audit unavailable");
-	});
 });
